requirements.txt
#TODO: Make better requirements, this is a direct dump of pip freeze

asgiref==3.2.3
Authlib==0.13
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
Django==2.2.9
django-appconf==1.0.3
django-braces==1.14.0
django-compressor==2.4
django-datatables-view==1.19.1
django-fsm==2.7.0
django-invitations==1.9.2
django-libsass==0.7
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.11.0
idna==2.8
libsass==0.19.4
oauthlib==3.1.0
psycopg2-binary==2.8.4
pycparser==2.19
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
rcssmin==1.0.6
requests==2.22.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.7